ReadMe
Needed to create a website that allows the user to add ratings on restaurants. Created Restaurant Model with appropriate fields so the admin can set the rating and fields accordingly. User is presented with a list of restaurants when application is launched. Used pagination to allow only 4 posts in one page. If user clicks on a restaurant link it renders the details of the restaurant. Added a dropdown to sort the list by highest rating. Use urls.py to map to views and render html pages accordingly. With additional time I could have added more navigation, fixed the footer to keep at the bottom, add user registration and login forms for security. Add maps of location of restaurants. Clean up the design to make it more appealing. Change the database to something more secure instead of sqlite. Fix rating so it works on all pages.  
Github at https://github.com/JT4life